[fei5585.1.hooksrefresh] Update docs and exports
