+++ __docs__/wonder-blocks-timing/use-interval.stories.tsx
@@ -1,8 +1,13 @@
 import * as React from "react";
 
 import {View} from "@khanacademy/wonder-blocks-core";
+import Button from "@khanacademy/wonder-blocks-button";
 
-import {useInterval} from "@khanacademy/wonder-blocks-timing";
+import {
+    useInterval,
+    ClearPolicy,
+    SchedulePolicy,
+} from "@khanacademy/wonder-blocks-timing";
 
 export default {
     title: "Packages / Timing / useInterval",
@@ -14,15 +19,43 @@ export default {
     },
 };
 
-export const BasicUsage = () => {
+export const Immediately = () => {
     const [callCount, setCallCount] = React.useState(0);
     const callback = React.useCallback(() => {
         setCallCount((callCount) => callCount + 1);
     }, []);
-    useInterval(callback, 1000);
+    const interval = useInterval(callback, 1000);
     return (
         <View>
+            <View>isSet = {interval.isSet.toString()}</View>
             <View>callCount = {callCount}</View>
+            <View style={{flexDirection: "row"}}>
+                <Button onClick={() => interval.set()}>Set interval</Button>
+                <Button onClick={() => interval.clear()}>Clear interval</Button>
+            </View>
+        </View>
+    );
+};
+
+export const OnDemandAndResolveOnClear = () => {
+    const [callCount, setCallCount] = React.useState(0);
+    const callback = React.useCallback(() => {
+        // eslint-disable-next-line no-console
+        console.log("action called");
+        setCallCount((callCount) => callCount + 1);
+    }, []);
+    const interval = useInterval(callback, 1000, {
+        clearPolicy: ClearPolicy.Resolve,
+        schedulePolicy: SchedulePolicy.OnDemand,
+    });
+    return (
+        <View>
+            <View>isSet = {interval.isSet.toString()}</View>
+            <View>callCount = {callCount}</View>
+            <View style={{flexDirection: "row"}}>
+                <Button onClick={() => interval.set()}>Set interval</Button>
+                <Button onClick={() => interval.clear()}>Clear interval</Button>
+            </View>
         </View>
     );
 };
