--- __docs__/wonder-blocks-timing/use-scheduled-interval.mdx
@@ -1,91 +1,0 @@
-import * as UseScheduledIntervalStories from './use-scheduled-interval.stories';
-
-import {Meta, Story, Canvas} from "@storybook/blocks";
-
-<Meta of={UseScheduledIntervalStories} />
-
-# `useScheduledInterval`
-
-`useScheduledInterval` is a hook that provides a convenient API for setting and clearing
-an interval. It is defined as follows:
-
-```ts
-function useScheduledInterval(
-    action: () => mixed,
-    timeoutMs: number,
-    options?: {|
-        schedulePolicy?: "schedule-immediately" | "schedule-on-demand",
-        clearPolicy?: "resolve-on-clear" | "cancel-on-clear",
-    |},
-): ITimeout;
-
-interface ITimeout {
-    get isSet(): boolean;
-    set(): void;
-    clear(policy?: ClearPolicy): void;
-}
-```
-
-By default the interval will be set immediately up creation. The `options` parameter can
-be used to control when when the interval is schedule and whether or not `action` should be
-called when the interval is cleared.
-
-Notes:
-
-* Because `clear` takes a param, it's import that you don't pass it directly to an event handler,
-  e.g. `<Button onClick={clear} />` will not work as expected.
-* Calling `set` after the interval has been cleared will restart the interval.
-* Updating the second paramter, `timeoutMs`, will also restart the interval.
-* When the component using this hooks is unmounted, the interval will automatically be cleared.
-* Calling `set` after the interval is already set does nothing.
-
-<Canvas>
-  <Story of={UseScheduledIntervalStories.Immediately} />
-</Canvas>
-
-```jsx
-const Immediately = () => {
-    const [callCount, setCallCount] = React.useState(0);
-    const callback = React.useCallback(() => {
-        setCallCount((callCount) => callCount + 1);
-    }, []);
-    const interval = useScheduledInterval(callback, 1000);
-    return (
-        <View>
-            <View>isSet = {interval.isSet.toString()}</View>
-            <View>callCount = {callCount}</View>
-            <View style={{flexDirection: "row"}}>
-                <Button onClick={() => interval.set()}>Set interval</Button>
-                <Button onClick={() => interval.clear()}>Clear interval</Button>
-            </View>
-        </View>
-    );
-};
-```
-
-<Canvas>
-  <Story of={UseScheduledIntervalStories.OnDemandAndResolveOnClear} />
-</Canvas>
-
-```jsx
-const OnDemandAndResolveOnClear = () => {
-    const [callCount, setCallCount] = React.useState(0);
-    const callback = React.useCallback(() => {
-        setCallCount((callCount) => callCount + 1);
-    }, []);
-    const interval = useScheduledInterval(callback, 1000, {
-        clearPolicy: ClearPolicy.Resolve,
-        schedulePolicy: SchedulePolicy.OnDemand,
-    });
-    return (
-        <View>
-            <View>isSet = {interval.isSet.toString()}</View>
-            <View>callCount = {callCount}</View>
-            <View style={{flexDirection: "row"}}>
-                <Button onClick={() => interval.set()}>Set interval</Button>
-                <Button onClick={() => interval.clear()}>Clear interval</Button>
-            </View>
-        </View>
-    );
-};
-```
