--- __docs__/wonder-blocks-timing/use-scheduled-timeout.mdx
@@ -1,92 +1,0 @@
-import * as UseScheduledTimeoutStories from './use-scheduled-timeout.stories';
-
-import {Meta, Story, Canvas} from "@storybook/blocks";
-
-<Meta of={UseScheduledTimeoutStories} />
-
-# `useScheduledTimeout`
-
-`useScheduledTimeout` is a hook that provides a convenient API for setting and clearing
-a timeout. It is defined as follows:
-
-```ts
-function useScheduledTimeout(
-    action: () => mixed,
-    timeoutMs: number,
-    options?: {|
-        schedulePolicy?: "schedule-immediately" | "schedule-on-demand",
-        clearPolicy?: "resolve-on-clear" | "cancel-on-clear",
-    |},
-): ITimeout;
-
-interface ITimeout {
-    get isSet(): boolean;
-    set(): void;
-    clear(policy?: ClearPolicy): void;
-}
-```
-
-By default the timeout will be set immediately up creation. The `options` parameter can
-be used to control when when the timeout is schedule and whether or not `action` should be
-called when the timeout is cleared.
-
-Notes:
-
-* Because `clear` takes a param, it's import that you don't pass it directly to an event handler,
-  e.g. `<Button onClick={clear} />` will not work as expected.
-* Calling `set` after the timeout has expired will restart the timeout.
-* Updating the second paramter, `timeoutMs`, will also restart the timeout.
-* When the component using this hooks is unmounted, the timeout will automatically be cleared.
-* Calling `set` after the timeout is set but before it expires means that the timeout will be
-  reset and will call `action`, `timeoutMs` after the most recent call to `set` was made.
-
-<Canvas>
-  <Story of={UseScheduledTimeoutStories.Immediately} />
-</Canvas>
-
-```jsx
-const Immediately = () => {
-    const [callCount, setCallCount] = React.useState(0);
-    const callback = React.useCallback(() => {
-        setCallCount((callCount) => callCount + 1);
-    }, []);
-    const {isSet, set, clear} = useScheduledTimeout(callback, 1000);
-    return (
-        <View>
-            <View>isSet = {isSet.toString()}</View>
-            <View>callCount = {callCount}</View>
-            <View style={{flexDirection: "row"}}>
-                <Button onClick={() => set()}>Set timeout</Button>
-                <Button onClick={() => clear()}>Clear timeout</Button>
-            </View>
-        </View>
-    );
-};
-```
-
-<Canvas>
-  <Story of={UseScheduledTimeoutStories.OnDemandAndResolveOnClear} />
-</Canvas>
-
-```jsx
-const OnDemandAndResolveOnClear = () => {
-    const [callCount, setCallCount] = React.useState(0);
-    const callback = React.useCallback(() => {
-        setCallCount((callCount) => callCount + 1);
-    }, []);
-    const {isSet, set, clear} = useScheduledTimeout(callback, 1000, {
-        clearPolicy: ClearPolicy.Resolve,
-        schedulePolicy: SchedulePolicy.OnDemand,
-    });
-    return (
-        <View>
-            <View>isSet = {isSet.toString()}</View>
-            <View>callCount = {callCount}</View>
-            <View style={{flexDirection: "row"}}>
-                <Button onClick={() => set()}>Set timeout</Button>
-                <Button onClick={() => clear()}>Clear timeout</Button>
-            </View>
-        </View>
-    );
-};
-```
