+++ __docs__/wonder-blocks-timing/use-timeout.mdx
@@ -6,44 +6,89 @@ import {Meta, Story, Canvas} from "@storybook/blocks";
 
 # `useTimeout`
 
-`useTimeout` is a hook that provides a simple API for using timers safely.
-It is defined as follows:
+`useTimeout` is a hook that provides a convenient API for setting and clearing
+a timeout. It is defined as follows:
 
 ```ts
 function useTimeout(
     action: () => mixed,
     timeoutMs: number,
-    active: boolean,
-): void;
+    options?: {|
+        schedulePolicy?: SchedulePolicy,
+        clearPolicy?: ClearPolicy,
+        actionPolicy?: ActionPolicy,
+    |},
+): ITimeout;
+
+interface ITimeout {
+    get isSet(): boolean;
+    set(): void;
+    clear(policy?: ClearPolicy): void;
+}
 ```
+
+By default the timeout will be set immediately up creation. The `options` parameter can
+be used to control when when the timeout is schedule and whether or not `action` should be
+called when the timeout is cleared.
 
 Notes:
 
-* Setting `active` to `true` will start the timeout and setting it to `false`
-  will stop it
-* Changing the value of `timeoutMs` will reset the timeout, changing `action`
-  will not.
+* Because `clear` takes a param, it's important that you don't pass it directly to an event handler,
+  e.g. `<Button onClick={clear} />` will not work as expected.
+* Calling `set` after the timeout has expired will restart the timeout.
+* Updating the second paramter, `timeoutMs`, will also restart the timeout.
+* When the component using this hooks is unmounted, the timeout will automatically be cleared.
+* Calling `set` after the timeout is set but before it expires means that the timeout will be
+  reset and will call `action`, `timeoutMs` after the most recent call to `set` was made.
 
 <Canvas>
-  <Story of={UseTimeoutStories.BasicUsage} />
+  <Story of={UseTimeoutStories.Immediately} />
 </Canvas>
 
 ```jsx
-export const BasicUsage = () => {
+const Immediately = () => {
     const [callCount, setCallCount] = React.useState(0);
-    const [active, setActive] = React.useState(false);
     const callback = React.useCallback(() => {
         setCallCount((callCount) => callCount + 1);
     }, []);
-    useTimeout(callback, 1000, active);
+    const {isSet, set, clear} = useTimeout(callback, 1000);
     return (
         <View>
+            <View>isSet = {isSet.toString()}</View>
             <View>callCount = {callCount}</View>
-            <View>active = {active.toString()}</View>
-            <Button onClick={() => setActive(!active)} style={{width: 200}}>
-                Toggle active
-            </Button>
+            <View style={{flexDirection: "row"}}>
+                <Button onClick={() => set()}>Set timeout</Button>
+                <Button onClick={() => clear()}>Clear timeout</Button>
+            </View>
         </View>
     );
 };
 ```
+
+<Canvas>
+  <Story of={UseTimeoutStories.OnDemandAndResolveOnClear} />
+</Canvas>
+
+```jsx
+const OnDemandAndResolveOnClear = () => {
+    const [callCount, setCallCount] = React.useState(0);
+    const callback = React.useCallback(() => {
+        setCallCount((callCount) => callCount + 1);
+    }, []);
+    const {isSet, set, clear} = useTimeout(callback, 1000, {
+        clearPolicy: ClearPolicy.Resolve,
+        schedulePolicy: SchedulePolicy.OnDemand,
+    });
+    return (
+        <View>
+            <View>isSet = {isSet.toString()}</View>
+            <View>callCount = {callCount}</View>
+            <View style={{flexDirection: "row"}}>
+                <Button onClick={() => set()}>Set timeout</Button>
+                <Button onClick={() => clear()}>Clear timeout</Button>
+            </View>
+        </View>
+    );
+};
+```
+./use-timeout.stories
