+++ packages/wonder-blocks-timing/src/index.ts
@@ -1,13 +1,3 @@
-import type {
-    IAnimationFrame,
-    IInterval,
-    IScheduleActions,
-    ITimeout,
-    WithActionScheduler,
-    WithActionSchedulerProps,
-    WithoutActionScheduler,
-} from "./util/types";
-
 export type {
     IAnimationFrame,
     IInterval,
@@ -16,12 +6,12 @@ export type {
     WithActionScheduler,
     WithActionSchedulerProps,
     WithoutActionScheduler,
-};
+    HookOptions,
+    Options,
+} from "./util/types";
 
-export {SchedulePolicy, ClearPolicy} from "./util/policies";
+export {SchedulePolicy, ClearPolicy, ActionPolicy} from "./util/policies";
 export {default as ActionSchedulerProvider} from "./components/action-scheduler-provider";
 export {default as withActionScheduler} from "./components/with-action-scheduler";
 export {useInterval} from "./hooks/use-interval";
 export {useTimeout} from "./hooks/use-timeout";
-export {useInterval as useScheduledInterval} from "./hooks/use-interval";
-export {useTimeout as useScheduledTimeout} from "./hooks/use-timeout";
